Refactoring. Fix errors in Ex1 and Ex2/
